feature-5: integrated react-router
