+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\App.js":"1","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\styles.js":"2","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\index.js":"3","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\index.js":"4"},{"size":752,"mtime":1606564222969,"results":"5","hashOfConfig":"6"},{"size":1006,"mtime":1606574194802,"results":"7","hashOfConfig":"6"},{"size":928,"mtime":1606574329640,"results":"8","hashOfConfig":"6"},{"size":197,"mtime":1606563306680,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1kkf05j",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\App.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\styles.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\index.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\index.js",[]]
+[{"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\styles.js":"1","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\index.js":"2","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\index.js":"3","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Search\\Search.js":"4","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Root.js":"5","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Search\\styles.js":"6","C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Form\\Form.js":"7"},{"size":1006,"mtime":1606574194802,"results":"8","hashOfConfig":"9"},{"size":937,"mtime":1606668616483,"results":"10","hashOfConfig":"9"},{"size":206,"mtime":1606600835632,"results":"11","hashOfConfig":"9"},{"size":942,"mtime":1606667111488,"results":"12","hashOfConfig":"9"},{"size":596,"mtime":1606668405890,"results":"13","hashOfConfig":"9"},{"size":128,"mtime":1606667349547,"results":"14","hashOfConfig":"9"},{"size":350,"mtime":1606668262993,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1kkf05j",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\styles.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\components\\Promotion\\Card\\index.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\index.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Search\\Search.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Root.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Search\\styles.js",[],"C:\\Users\\Samir\\Desktop\\devsamir\\projetos-a-vera\\gatry-clone\\src\\pages\\Promotion\\Form\\Form.js",[]]
