+++ src/components/Promotion/Card/index.js
@@ -5,7 +5,7 @@ import { Container, CardImg, CardInfo, CardComment, CommentCounter, CardLink } f
 function PromotionCard({ promotion }) {
   return (
       <Container>
-          <CardImg src={promotion.imageUrl} alt="imagem" />
+          <CardImg src={promotion.imageUrl} alt={promotion.title}/>
           <CardInfo>
             <h1>{promotion.title}</h1>
             <span>{promotion.price}</span>
@@ -18,7 +18,7 @@ function PromotionCard({ promotion }) {
                 {promotion.comments.length}{' '}
                 {promotion.comments.length > 1 ? 'Comentários' : 'Comentário'}
               </CommentCounter>
-              <CardLink href={promotion.url} target="_blank">Ir para o site</CardLink>
+              <CardLink href={promotion.url} target="_blank" >Ir para o site</CardLink>
             </footer>
           </CardInfo>
       </Container>
