+++ src/pages/Promotion/Search/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  max-width: 800px;
+  margin: 30px auto;
+`;
